requirements.txt
@'
requests
feedparser
python-dateutil
jinja2
'@ | Set-Content .\requirements.txt -Encoding utf8
